@@ -3,9 +3,10 @@
     public class Constants
     {
         #region [ Attributes ]
-        const string _urlApi = "https://myvetdacalo.azurewebsites.net/";
+        //const string URL_BASE = "https://myvetdacalo.azurewebsites.net/";
+        const string _urlBase = "http://192.168.0.194:45455/";
         const string _prefix = "api/";
-        public static string URL_API = _urlApi;
+        public static string URL_BASE = _urlBase;
         public static string PREFIX = _prefix;
 
         public static string LicenseSyncfusion = "";
