@@ -89,7 +89,7 @@ namespace MyVet.Prism.ViewModels
             };
 
             var response = await _apiService.RegisterUserAsync(
-                Constants.URL_API,
+                Constants.URL_BASE,
                 Constants.PREFIX,
                 "Account",
                 request);
