@@ -54,7 +54,7 @@ namespace MyVet.Prism.ViewModels
             };
 
             var response = await _apiService.RecoverPasswordAsync(
-                Constants.URL_API,
+                Constants.URL_BASE,
                 Constants.PREFIX,
                 "Account/RecoverPassword",
                 request);
