@@ -28,7 +28,7 @@ namespace MyVet.Web.Controllers
 
         public IActionResult Index()
         {
-            return new ViewAsPdf(_dataContext.Pets
+            return View(_dataContext.Pets
                 .Include(p => p.Owner)
                 .ThenInclude(o => o.User)
                 .Include(p => p.PetType)
