@@ -55,7 +55,7 @@
 
 <hr />
 <div class="row">
-    <div class="col-md-1">
+    <div class="col-md-2">
         <a asp-action="MyPets" class="btn btn-primary">Sus mascotas</a>
     </div>
     <div class="col-md-2">
