@@ -1,13 +1,13 @@
 ﻿@model MyVet.Web.Data.Entities.Manager
 
 @{
-    ViewData["Title"] = "Deatalles";
+    ViewData["Title"] = "Detalles";
 }
 
 <h2>Administrador</h2>
 
 <div>
-    <h4>ViewData["Title"]</h4>
+    <h4>Detalles</h4>
     <hr />
     <dl class="dl-horizontal">
         <dt>
