@@ -3,7 +3,7 @@
     ViewData["Title"] = "Editar";
 }
 
-<h2>ViewData["Title"]</h2>
+<h2>Editar</h2>
 
 <h4>Administrador</h4>
 <hr />
