@@ -4,7 +4,7 @@
     ViewData["Title"] = "Crear";
 }
 
-<h2>ViewData["Title"]</h2>
+<h2>Crear</h2>
 
 <h4>Mascota</h4>
 <hr />
