@@ -7,7 +7,7 @@
 <h2>Propietario</h2>
 
 <div>
-    <h2>ViewData["Title"]</h2>
+    <h2>Detalles</h2>
     <hr />
     <dl class="dl-horizontal">
         <dt>
@@ -50,13 +50,13 @@
 </div>
 <div>
     <a asp-action="Edit" asp-route-id="@Model.Id" class="btn btn-warning">Editar</a>
-    <a asp-action="AddPet" asp-route-id="@Model.Id" class="btn btn-primary">Add Pet</a>
+    <a asp-action="AddPet" asp-route-id="@Model.Id" class="btn btn-primary">Agregar Mascota</a>
     <a asp-action="Index" class="btn btn-success">Regresar a la lista</a>
 </div>
-<h4>Pets</h4>
+<h4>Mascotas</h4>
 @if (Model.Pets.Count == 0)
 {
-    <h5>Not pets added yet.</h5>
+    <h5>No se han agregado mascotas todavía.</h5>
 }
 else
 {
@@ -104,9 +104,9 @@ else
                         @Html.DisplayFor(modelItem => item.Remarks)
                     </td>
                     <td>
-                        <a asp-action="EditPet" asp-route-id="@item.Id" class="btn btn-warning">Editar</a>
-                        <a asp-action="DetailsPet" asp-route-id="@item.Id" class="btn btn-info">Detalles</a>
-                        <button data-id="@item.Id" class="btn btn-danger deleteItem" data-toggle="modal" data-target="#deleteDialog">Eliminar</button>
+                        <a asp-action="EditPet" asp-route-id="@item.Id" class="btn btn-warning"><i class="glyphicon glyphicon-pencil"></i></a>
+                        <a asp-action="DetailsPet" asp-route-id="@item.Id" class="btn btn-info"><i class="glyphicon glyphicon-list"> </i></a>
+                        <button data-id="@item.Id" class="btn btn-danger deleteItem" data-toggle="modal" data-target="#deleteDialog"><i class="glyphicon glyphicon-trash"></i></button>
                     </td>
                 </tr>
             }
